@@ -6,6 +6,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
+  "type": "module",
   "keywords": [
     "nodejs",
     "express",
@@ -38,7 +39,7 @@
     "sequelize": "^5.22.5"
   },
   "devDependencies": {
-    "eslint": "^8.10.0",
-    "eslint-plugin-react": "^7.29.3"
+    "eslint": "^8.11.0",
+    "eslint-plugin-react": "^7.29.4"
   }
 }
